backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
openai==1.3.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
numpy>=1.26.4
